+++ src/main/java/racingcar/view/InputView.java
@@ -4,8 +4,8 @@ import camp.nextstep.edu.missionutils.Console;
 
 import java.util.Set;
 
-import static racingcar.domain.InputCarValidator.carsByStringToArray;
-import static racingcar.domain.InputCarValidator.checkDuplicateCarName;
+import static racingcar.domain.CarValidator.carsByStringToArray;
+import static racingcar.domain.CarValidator.checkDuplicateCarName;
 
 public class InputView {
     private static final String INPUT_CAR_NAMES = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
